requirements.txt
aiogram==3.7.0
pydantic-settings==2.2.1
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.0b2
apscheduler==3.10.4
azure-cognitiveservices-speech==1.41.1
ffmpeg==1.4





